test(prover-node): add tx-proof revival case to prover restart e2e
Splits the prover-restart e2e into two cases (now a .parallel file):
- in-flight top-tree jobs: gates the top tree so its jobs are in flight at a
  clean shutdown, and asserts they are preserved (zero cancels) and reused on
  restart. Discriminates the clean-shutdown abortJobs:false fix.
- in-flight transaction proofs: starves agents from the start so real tx
  base-rollup proofs sit unproven at the broker, then asserts those non-parity
  proofs survive a clean shutdown and are re-requested/reused on restart rather
  than the leaf parity jobs the top-tree case would otherwise dominate.
